pyproject.toml
[tool.poetry]
name = "sonofflan"
version = "0.1.0"
description = "Library to interact with Sonoff devices through LAN mode (without eWeLink cloud)"
authors = ["Danilo Treffiletti <urban82@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
keywords = ["sonoff", "ewelink", "itead"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Networking",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.7"
pycryptodome = "^3.15.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
